refactor: refine home screen state management and polish UI visual consistency
- Implement `HomeStateManager` and `HomeUiState` to centralize Home screen logic, including splash and content state transitions.
- Introduce `HomeSplashLoader` to provide an animated icon-cycling sequence during the initial app load using `AnimatedContent` and shared elements.
- Refactor `GenreMemoriesLoader` to use `AnimatedContent` for smoother transitions and implement explicit image loading checks via Coil.
- Standardize iconography across `ObjectiveOverlay`, `AdvanceTrigger`, and `NewSagaLibrary` by transitioning to `themePainter()` and `sagaBrush()` utilities.
- Redesign the "Universe Echoes" section in `NewSagaLibrary` into a grid of levitating `EchoIdeaCard` components with improved shadow and border styling.
- Simplify theme configuration by migrating `cornerSizeDp` to an integer-based system in `GenreVisualConfig` and centralizing resolution logic.
- Optimize global navigation transitions in `MainActivity` with unified `SAGA_THEME_TRANSITION_MS` timing and `FastOutSlowInEasing` curves.
- Refine `SagaIdeationService` by adjusting AI generation temperature and model requirements for more consistent universe suggestions.
- Clean up redundant `sharedElement` transitions in `HomeView`, `SettingsView`, and `NewSagaView` to improve layout performance.
- Update UI components to use `surfaceContainer` and `labelLarge` typography for better adherence to Material 3 standards.

+++ app/src/main/java/com/ilustris/sagai/core/ai/model/GenreVisualConfig.kt
@@ -26,7 +26,7 @@ data class GenreVisualConfig(
     val colorPalette: List<String> = emptyList(),
     // ── Shape ────────────────────────────────────────────────────────
     @SerializedName("cornerSizeDp")
-    val cornerSizeDp: Float = 0f,
+    val cornerSizeDp: Int = 0,
     // ── Haptics (waveform in ms, tuned per genre SFX — see Remote Config) ──
     @SerializedName("vibrationPattern")
     val vibrationPattern: List<Long> = emptyList(),
